fix comment id bug
